+++ crates/lib/src/core/parser/grammar/base.rs
@@ -21,7 +21,7 @@ pub struct Ref {
     reset_terminators: bool,
     allow_gaps: bool,
     optional: bool,
-    cache_key: u32,
+    cache_key: u64,
     simple_cache: OnceLock<Option<(AHashSet<String>, AHashSet<&'static str>)>>,
 }
 
@@ -140,14 +140,14 @@ impl Matchable for Ref {
         })
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 }
 
 #[derive(Clone, Debug)]
 pub struct Anything {
-    cache_key: u32,
+    cache_key: u64,
     terminators: Vec<Arc<dyn Matchable>>,
 }
 
@@ -178,7 +178,7 @@ impl Anything {
 impl Segment for Anything {}
 
 impl Matchable for Anything {
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
