+++ crates/lib/src/core/parser/parsers.rs
@@ -16,7 +16,7 @@ pub struct TypedParser {
     instance_types: Vec<String>,
     optional: bool,
     trim_chars: Option<Vec<char>>,
-    cache_key: u32,
+    cache_key: u64,
     factory: fn(&dyn Segment) -> ErasedSegment,
 }
 
@@ -54,7 +54,7 @@ impl TypedParser {
 impl Segment for TypedParser {}
 
 impl Matchable for TypedParser {
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 
@@ -95,7 +95,7 @@ pub struct StringParser {
     type_: Option<String>,
     optional: bool,
     trim_chars: Option<Vec<char>>,
-    cache_key: u32,
+    cache_key: u64,
 }
 
 impl StringParser {
@@ -164,7 +164,7 @@ impl Matchable for StringParser {
         Ok(MatchResult::empty_at(idx))
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 }
@@ -174,7 +174,7 @@ pub struct RegexParser {
     template: Regex,
     anti_template: Option<Regex>,
     factory: fn(&dyn Segment) -> ErasedSegment,
-    cache_key: u32,
+    cache_key: u64,
 }
 
 impl PartialEq for RegexParser {
@@ -255,7 +255,7 @@ impl Matchable for RegexParser {
         Ok(MatchResult::empty_at(idx))
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 }
@@ -265,7 +265,7 @@ pub struct MultiStringParser {
     templates: AHashSet<String>,
     simple: AHashSet<String>,
     factory: fn(&dyn Segment) -> ErasedSegment,
-    cache: u32,
+    cache: u64,
 }
 
 impl MultiStringParser {
@@ -326,7 +326,7 @@ impl Matchable for MultiStringParser {
         Ok(MatchResult::empty_at(idx))
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache
     }
 }
